Revert "[Volume 7] 이벤트 기반 아키텍처 및 Kafka 파이프라인 구축"

+++ apps/commerce-api/src/main/java/com/loopers/interfaces/coupon/dto/AdminCouponV1Dto.java
@@ -52,7 +52,7 @@ public class AdminCouponV1Dto {
         long discountValue,
         Long minOrderAmount,
         int totalQuantity,
-        long issuedQuantity,
+        int issuedQuantity,
         ZonedDateTime expiredAt,
         ZonedDateTime createdAt,
         ZonedDateTime updatedAt
@@ -72,7 +72,7 @@ public class AdminCouponV1Dto {
         String discountType,
         long discountValue,
         int totalQuantity,
-        long issuedQuantity,
+        int issuedQuantity,
         ZonedDateTime expiredAt,
         ZonedDateTime createdAt
     ) {
